requirements.txt
openai-whisper
sounddevice
numpy
scipy
torch
transformers
